+++ src/Blameable/Mapping/Driver/Yaml.php
@@ -37,6 +37,8 @@ class Yaml extends File implements Driver
         'one',
         'string',
         'int',
+        'ulid',
+        'uuid',
     ];
 
     /**